requirements.txt
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
torch
transformers
accelerate
requests
dotenv
python-telegram-bot
langchain-google-genai
google-generativeai
